Custom chef bootstrap template.
This template is based of the default ubuntu10.04-gems template
with the addition of creating a special user for use by the
chef server.

The user's name is chef and password is currently set as
password (use mkpasswd on ubuntu to generate a new password
and replace it in this template).