Addition of the community apt cookbook.
This cookbook includes recipes to execute apt-get update
to ensure the local APT package cache is up to date. There
are recipes for managing the apt-cacher-ng caching proxy and
proxy clients. It also includes a LWRP for managing APT
repositories in /etc/apt/sources.list.d as well as an LWRP
for pinning packages via /etc/apt/preferences.d.